// src/client/src/app/models/PlayerStatModel.model.ts
export class PlayerStatModel {
    PlayerID: number = 0;
    Points: number = 0.0;
    GamesPlayed: number = 0;
    GamesWon: number = 0;
    TotalNumberOfPartners: number = 0;
    TotalNumberOfOpponents: number = 0;
    AveragePlayerLevelOfPartners: number = 0.0;
    AveragePlayerLevelOfOpponents: number = 0.0;
}